backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
dateutils==0.6.12
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-cors-headers==4.2.0
django-import-export==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
psycopg2==2.9.6
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
sendgrid==6.10.0
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
tablib==3.5.0
tzdata==2023.3
urllib3==2.0.4
xlrd==2.0.1
xlwt==1.3.0
